move tasks back and fix digest func
