+++ superset/models/slice.py
@@ -42,9 +42,9 @@ from sqlalchemy.orm.mapper import Mapper
 from superset import db, is_feature_enabled, security_manager
 from superset.legacy import update_time_range
 from superset.models.helpers import AuditMixinNullable, ImportExportMixin
+from superset.tasks.thumbnails import cache_chart_thumbnail
 from superset.tasks.utils import get_current_user
 from superset.thumbnails.digest import get_chart_digest
-from superset.thumbnails.tasks import cache_chart_thumbnail
 from superset.utils import core as utils
 from superset.utils.memoized import memoized
 from superset.viz import BaseViz, viz_types
